requirements.txt
langchain-groq
langchain-huggingface
faiss-cpu
python-dotenv
streamlit
langchain
langchain-community

cryptography
pymupdf

hf_xet  # for downloading hugging face 